atividade_fab/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
motor>=3.4
python-dotenv>=1.0
pika
fastapi
uvicorn
python-dotenv
